[UPD] Massive eslint fixes, TS and folder updates
