--- app/components/Menu/Menu.scss
@@ -1,58 +1,0 @@
-@import '../../assets/mixins',
-        '../../assets/vars';
-
-.menu {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-
-    &__item {
-        margin: 0;
-        padding: 7px 10px;
-        display: inline-block;
-        line-height: 1;
-        cursor: pointer;        
-        border-radius: 5px;
-
-        &:hover {
-            background: $hover-reaction;
-
-            .menu--sub {
-                display: block;
-            }
-        }
-    }
-
-    &--horizontal {
-        > .menu-item {
-            margin: 0 6px;
-        }
-    }
-
-    &--vertical {
-        .menu-item {
-            width: 100%;
-        }
-    }
-
-    &--sub {
-        display: none;
-        position: absolute;
-        background: $background-dark-color;
-        max-width: 240px;
-        font-size: 12px;
-        margin: 8px 0 0 -10px;
-        z-index: 3;
-
-        .menu-item {
-            border: none;
-            padding: 10px;
-        }
-    }
-}
-
-.menu-item {
-    &__right {
-        float: right;
-    }
-}
