+++ app/components/module/main/StatusBar/StatusBar.scss
@@ -1,0 +1,13 @@
+@import '../../assets/mixins', '../../assets/vars';
+
+.status-bar {
+  background: $background-dark-color;
+  // @include flex(row, flex-start, center);
+
+  font-size: 12px;
+
+  &__item {
+    padding: 5px 12px;
+    border-right: 1px solid $border-color;
+  }
+}
