--- app/containers/App/AppConnect.js
@@ -1,19 +1,0 @@
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-import App from './index';
-import * as AppActions from '../../actions/app';
-
-function mapStateToProps(state) {
-  return {
-    settingsVisible: state.appMain.settingsWinVisible
-  };
-}
-
-function mapDispatchToProps(dispatch) {
-  return bindActionCreators(AppActions, dispatch);
-}
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(App);
