--- app/containers/Settings/SettingsConnect.js
@@ -1,10 +1,0 @@
-import { connect } from 'react-redux';
-import Settings from './index';
-
-function mapStateToProps(state) {
-  return {
-    winVisible: state.appMain.settingsWinVisible
-  };
-}
-
-export default connect(mapStateToProps)(Settings);
