+++ app/data/common/paths.js
@@ -1,0 +1,10 @@
+const paths = {
+  appData: `${process.env.APPDATA}/Octagon Modmaking Studio/Data`,
+  octagonCfg: '/octagon.omsdata',
+  langsFolder: '/Languages',
+  session: '/session.omsdata',
+  theme: '/theme.omsdata',
+  user: '/user.omsdata',
+};
+
+export default paths;
