--- app/data/paths.js
@@ -1,10 +1,0 @@
-const paths = {
-  appData: `${process.env.APPDATA}/Octagon Modmaking Studio/Data`,
-  octagonCfg: '/octagon.omsdata',
-  langsFolder: '/Languages',
-  session: '/session.omsdata',
-  theme: '/theme.omsdata',
-  user: '/user.omsdata'
-};
-
-export default paths;
