+++ app/modules/main/MainMenu/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { Menu } from 'antd';
+
+export default class MainMenu extends React.Component {
+  public render() {
+    return <Menu mode="horizontal" className="main-win__top-nav top-nav" />;
+  }
+}
